+++ dash_auth/__init__.py
@@ -1,5 +1,6 @@
+from .public_routes import add_public_routes, public_callback
 from .basic_auth import BasicAuth
 from .version import __version__
 
 
-__all__ = ["BasicAuth", "__version__"]
+__all__ = ["add_public_routes", "public_callback", "BasicAuth", "__version__"]